Adding project file and moving /unlol
Also fixed Vault dep
